prepare frontend for production

+++ client/src/components/messenger/info/ChatInfo.tsx
@@ -5,7 +5,7 @@ import useChat from "../../../store/useChat"
 import { AuthState } from "../../../context/useAuth"
 import { chatService } from "../../../services/chat.service"
 interface Props {
-      conversationUser: User
+      conversationUser: User | undefined
 }
 
 export default function ChatInfo ({ conversationUser }: Props) {
@@ -14,30 +14,28 @@ export default function ChatInfo ({ conversationUser }: Props) {
       const { user } = AuthState()
 
       async function onRemoveChat () {
-            const chat = chats.find(chat => chat._id === selectedChat._id)
-
+            const chat = chats.find(chat => chat._id === selectedChat?._id)
+            if (!chat || !user) return
             await chatService.removeChat(chat._id, user._id)
-            const newChats = chats.filter(chat => chat._id !== selectedChat._id)
+            const newChats = chats.filter(chat => chat._id !== selectedChat?._id)
             setChats(newChats)
             setSelectedChat(null)
-
-
       }
 
       return (
             <section className="w-full h-full">
                   <div className="flex flex-col items-center h-full">
                         <div className="flex flex-col items-center">
-                              <img src={conversationUser.profileImg} className="w-24 h-24 md:w-32 md:h-32 rounded-full shadow-lg mb-5 object-cover object-top" alt="profile" />
-                              <span className="text-2xl font-semibold">{conversationUser.username}</span>
-                              <span className="text-gray-500">{conversationUser.email}</span>
+                              <img src={conversationUser?.profileImg} className="w-24 h-24 md:w-32 md:h-32 rounded-full shadow-lg mb-5 object-cover object-top" alt="profile" />
+                              <span className="text-2xl font-semibold">{conversationUser?.username}</span>
+                              <span className="text-gray-500">{conversationUser?.email}</span>
                         </div>
 
                         <div className="[&>*]:border-y-[6px] flex flex-col  h-full gap-y-6 w-full [&>*]:border-gray-200 [&>*]:px-6">
                               <div className="mt-5">
                                     <div className="flex flex-col pt-2">
                                           <h2 className="text-lg">About</h2>
-                                          <p className="text-gray-500 py-2 text-sm">{conversationUser.about || 'Hello There :) '}</p>
+                                          <p className="text-gray-500 py-2 text-sm">{conversationUser?.about || 'Hello There :) '}</p>
                                     </div>
                               </div>
